Fixed some compiler warnings and minor code tidy up.

+++ src/PropertyForm.cpp
@@ -610,7 +610,7 @@ BOOL CPropertyForm::PreTranslateMessage(MSG *pMsg)
 	if (pMsg && pMsg->message == WM_KEYDOWN) {
 		if (pMsg->wParam == VK_TAB && container) {
 			if ((0x8000 & GetKeyState(VK_CONTROL)) && !(0x8000 & GetKeyState(VK_MENU))) {	// control-tab to change page
-				const int numPages = container->pages.GetCount();
+				const int numPages = static_cast<int>(container->pages.GetCount());
 				if (numPages > 1) {
 					const int increment = (0x8000 & GetKeyState(VK_SHIFT)) ? -1 : 1;		// added shift iterates backwards
 					int newPage = tabs.GetCurSel() + increment;
